add check for data type input
