+++ src/scripts/dbgap/create_dbgap_xml_files.py
@@ -1,5 +1,6 @@
 import argparse
 import sys
+sys.path.append("./")
 from src.services.terra import TerraAPIWrapper
 from dbgap_classes import Sample, ReadGroup, Experiment, Run, Submission
